Got SPI configuration working; Double check var and config
Don & I met in the RF lab for 2.5 hours (7pm-9:30pm) to work on the SPI protocol. After tweaking the functions and double checking the logic we realized that the pins used for SPI were not ever configured. We configured them and another issue we encounter is that the CS was constantly being changed during the read/write state with messed with the data that was being acquired from the peripheral. 

Set up the flag var to static so that once the time is set then the clock will keep going due to the battery. Additionally, double check SPI config to really understand why those bits are set and document the process in code.