display_time func with seconds as a demo for getting time
Bradley, Kym, and Harrison met in the RF lab from 11:20 am - 1:30 pm to work on getting the buttons working. A basic prototype was wired and tested and the ISR should be functional and just needs to now be implemented to change time. Another button needs to be configured to handle a write to initially configure the timer so that it is not overwritten everytime the power cycles